[perf_automation] Rename tracy.py -> tracy_tool.py to avoid shadowing the tracy module
